[MLIR] Introduce placeholder (m_Placholder)
Introduce data structures for placeholder. It is a centralized matching around
the matching context class, which has a "global" view on what is going on
during the matching phase. When a placeholder is instatiated it is also
registered to the context. During the matching a placeholder asks the context
if he can match the current Value.