Make symbolTableMap template to store a string or a Tensor
Initially, in the symbolTableMap, we were storing only the array identifier.
While this is sufficient for the builder logic, it is not for the matcher logic
as we need to distinguish accesses to the same array but with different access
patterns (i.e., Syrk access pattern).  Thus, we extended the symbol Table to
store a tensor type. A tensor type is identified by its name as well as its
induction variables.