[MLIR] Fix MatvecOp and MatmulOp after last rebase
Commit eae76faeeae7110389047c9d1ebe90990dec3497 starts to retire linalg C++
implementation in favour of a linalg-ods-gen'd op. But the new flow
(`mlir-linalg-ods-gen`) does not allow to specify constant (i.e., alpha and
beta). If we want to extend the new flow to allows for constant, we need to
extend the parsing to multiple comprehensions. For now, we still rely on the
old C++-based flow.