[MLIR] Allow placeholders overloading in Access Matcher
Until now the access matchers were able to capture any affine expression of
the form: coefficient * induction + step. This was a limitation if we want to
match more complex access patterns such as convolution. This is a preliminary
patch to allow placeholder overloading (i.e., _i + _j) and thus enable matching
convolutions.  A placeholder overloading creates another placeholder which has
two ids; a better way would be to group the placeholders in a placeholderGroup.